added tests for checking deploy status
