+++ src/lib/helpers.ts
@@ -155,7 +155,7 @@ export async function namespaceDeploymentsReady(namespace: string = "pepr-system
   while (!ready) {
     ready = await checkDeploymentStatus(namespace);
     if (ready) {
-      break;
+      return ready;
     }
     await new Promise(resolve => setTimeout(resolve, 1000));
   }
